+++ docsource/CHANGELOG.md
@@ -11,6 +11,7 @@ See [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) for more details.
 - Doc: CESMDV2 incident documentation 
 - load CESMDV2 record
 - test: CESMDV2 read file
+- plot stations on basemap
 
 
 ### Changed